issue #88 DateFragment DataBinding
